+++ src-ui/app/views/dashboard-view/views/brightness-automations-view/brightness-automations-view.component.ts
@@ -1,48 +1,18 @@
 import { Component } from '@angular/core';
-import { ModalService } from '../../../../services/modal.service';
-import {
-  ConfirmModalComponent,
-  ConfirmModalInputModel,
-  ConfirmModalOutputModel,
-} from '../../../../components/confirm-modal/confirm-modal.component';
-import { hshrink } from '../../../../utils/animations';
 import { AutomationConfigService } from '../../../../services/automation-config.service';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
-import { BrightnessAutomationsConfig } from '../../../../models/automations';
 
 @Component({
   selector: 'app-brightness-automations-view',
   templateUrl: './brightness-automations-view.component.html',
   styleUrls: ['./brightness-automations-view.component.scss'],
-  animations: [hshrink()],
 })
 export class BrightnessAutomationsViewComponent {
   advancedMode = false;
 
-  constructor(
-    private modalService: ModalService,
-    private automationConfigService: AutomationConfigService
-  ) {
+  constructor(private automationConfigService: AutomationConfigService) {
     this.automationConfigService.configs.pipe(takeUntilDestroyed()).subscribe((configs) => {
       this.advancedMode = configs.BRIGHTNESS_AUTOMATIONS.advancedMode;
     });
-  }
-
-  showModeInfoModal() {
-    this.modalService
-      .addModal<ConfirmModalInputModel, ConfirmModalOutputModel>(ConfirmModalComponent, {
-        title: 'brightness-automations.modeInfoModal.title',
-        message: 'brightness-automations.modeInfoModal.message',
-        confirmButtonText: 'shared.modals.ok',
-        showCancel: false,
-      })
-      .subscribe();
-  }
-
-  async toggleAdvancedMode() {
-    await this.automationConfigService.updateAutomationConfig<BrightnessAutomationsConfig>(
-      'BRIGHTNESS_AUTOMATIONS',
-      { advancedMode: !this.advancedMode }
-    );
   }
 }