+++ src-ui/app/views/dashboard-view/views/shutdown-automations-view/tabs/shutdown-automations-triggers-tab/shutdown-automations-triggers-tab.component.ts
@@ -168,6 +168,16 @@ export class ShutdownAutomationsTriggersTabComponent implements OnInit {
     );
   }
 
+  async toggleTriggerWhenAloneOnlyWhenSleepModeActive() {
+    await this.automationConfigs.updateAutomationConfig<ShutdownAutomationsConfig>(
+      'SHUTDOWN_AUTOMATIONS',
+      {
+        triggerWhenAloneOnlyWhenSleepModeActive:
+          !this.config.triggerWhenAloneOnlyWhenSleepModeActive,
+      }
+    );
+  }
+
   async toggleWhenAloneActivationWindow() {
     // Toggle the activation window
     const config: Partial<ShutdownAutomationsConfig> = {